Reparameterise OU from beta/sigma to tau/kappa
Change parameterisation of OU process model. It used to be specified in
terms of beta (reversion parameter) and sigma (diffusion parameter), and
it is now specified by tau (time scale of autocorrelation) and kappa
(long-term variance). This new parameterisation is more numerically stable
as it circumvents the identifiability issues of the standard OU
parameters.