Remove @details from method documentation
Using the "@details" keyword in the documentation of R6 methods
created entries at the top of the documentation (for the whole
class rather than just the specific method). Removing the
keyword seems to have solved the problem.

This is surprising because I seem to remember having the exact
opposite problem a while ago (which is why I included the
@details keyword in the first place).