Add constant to likelihood penalty
The penalty of the likelihood function is a normal pdf,
and we had omitted a couple of terms which do not depend
on the parameters. Those terms might be important in
some situations, where we need to know the actual value
of the likelihood (rather than up to a multiplicative
constant), e.g. for model comparison.