Allow for non-zero velocity in CTCRW
I had to change a few things to allow for non-zero mean velocity in
the CTCRW model. Changes to the R code were minimal (mostly just
adding the parameter entry for the mean to the link and inverse
link function lists). In the TMB likelihood code, I added a term
to the state update in the Kalman filter, representing the added
terms in the transition densities of the velocity and location
processes. I think that the additional terms can be viewed as
the "control" part of the Kalman filter.