Reparameterise CTCRW in terms of beta and nu
The CTCRW model was parameterised in terms of beta and sigma,
the reversion parameter and diffusion parameter of the
underlying OU process. That specification is the most obvious
one because it's based on the SDE parameters, but it seems to
be susceptible to identifiability issues. The parameters are
also hard to interpret. For these reasons, I changed to the
parameters tau (1/beta, time scale of autocorrelation) and
nu (sqrt(pi/beta)*sigma/2, mean speed), which are more readily
interpretable and will hopefully be more identifiable.