Change format of CI methods to array
The methds SDE$CI_pointwise and SDE$CI_simultaneous used
to return a list with two elements (low and upp). This was
not very convenient to use and to quickly see the CIs of
constant parameters. So, I've changed it so they now
return arrays with one row for each parameter, one column
for each bound (upper and lower), and one layer for each
time step. By default, it just shows the CIs for the
first observation (useful for constant parameters).