Fix effective degrees of freedom for random effects
There was still a couple of issues with the way the EDFs were computed (for
use in conditional AIC), namely: (1) the joint log-likelihood used in the conditional AIC should not include the penalty but just the data likelihood, and (2) the I matrix in Wood et al. (2016, JASA) is the Hessian of the log-likelihood (in our case, that's not XtX) which is not computed from joint likelihood TMB object.