Remove do.call for optim call
This replaces do.call(optim, ...) by optim(...), thus making the arguments explicit (par, fn, gr). This small change fixed a problem I had with the CIR model, where the optimisation resulted in a "memory not mapped" error. Not clear why, but it might be that do.call was passing an undesirable argument to optim.